+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.5.3] - 2023-05-18
+
+### Fixed
+
+- Process memory scanning on 64-bit, cooldowns should be more reliable now
+- Reset cooldown pointer on re-log same client instance
+
 ## [4.5.2] - 2023-05-18
 
 ### Added